Add dynamic inputs to net, reshuffle some init code
In main.c add an optional arg for input vectors.  Each vector
is shown for the same number of timesteps.  This was the simplest
incremental change but there are probably better ways of doing
this.

For many benchmarks, the user will present a series of inputs
e.g. a set of MNIST digits, a number of audio windows, in
a single application simulation.  For my kernel, this equates
to running a number of timesteps, changing the network, then
running another set of timesteps.  Basically I just need a
way to feed in new inputs / parameters mid simulation.
A better way might be to have some commands e.g. a run command
and a set command to change values.