Parameterise the chip using a separate technology file
Ultimately we want to make the simulation configurable, so we can try
different architectures. Make all components e.g. number of cores,
compartments, fan outs and energy / time metrics configurable. This
uses a separate file with a simple format i.e.

<metric name> <metric value>
[Each line is a different parameter.]
As an example see loihi.tech

This is probably over simplified, but it's a step in the direction,
allowing the user to define the chip dynamically. In theory we should be
able to use those parameters for another digital chip e.g. Loihi 2
or TrueNorth. Next I need to see what kind of different decisions were
made for mixed-signal designs

The only painful thing here is a lot of simple arrays now need to
be dynamically allocated. Allocation of the chip memory now looks
pretty convoluted