[WIP] Iterating towards next design, code now builds
I'm using a common format to input both spiking networks and the
architecture (as a set of commands). It isn't important but it
just seems to work. Those commands will be replaced by Python calls
to an API (when using Python as glue logic).

Network and architecture are reasonably separated. I still need
to add functions to map one to the other. I still need to
un-break the parsing of the architecture features e.g. somas,
axons to get the whole thing running again.

I also need to add the function that maps the tiles to the
NoC interconnect. If we have a command / API to map, we could do
crazy stuff like change the network links mid-simulation. That
could be useful for some really interesting experiments.