Continue reworking python interface
Improve scripts to make it easier to create the simulation
programs. The simulator was also changed to map each neuron
to hardware rather than neuron groups. Neuron groups are just
there to share network parameters. I'm separating the
experiments into the experimental setup and some generic
library functions, like creating a layer of neurons.
Using Python classes seems to make sense. We define a network
then we can save the structure to a file in a format that
the simulator can read.

A neuron group can span over multiple cores. This was to
massively simplify things like networks with layers. A
single layer can then be a neuron group.

The main thing I don't like is it feels like we're implementing
the same structures twice, once in the Python API and once
in the C framework.