Experiment with timing model, considering pipeline
Calibrate according to measured data. It looks like we were
missing a couple of metrics as well. With these the
simulations match up better. However, I need to consider
backpressure from the synaptic units, this limits the
parallelism and when we model this the numbers match
up better. This model is really a hack though and I
need to better consider the timing for more complex
networks to be simulated accurately. I think I
understand what's happening now though.