Minor reverts and changes to script params
It seems this time that the input spikes were one off the other
way. Maybe this is to do with how the timing probe is working.
Maybe it isn't reliably sampling the first timestep?
If so, we can just adjust the data set in the compress scripts
i.e. my code is fine