Redo how we store SNN connectivity using axons
I want to store information about incoming connections at
the receiving core. This will allow me to restructure the
code to be more easily parallelized. I want to store
information about incoming connections at the receiving core
i.e., if a spike is received at a core, which connections
will be read on the message receiving pass.

The mapping of connections to axons is working reliably. It
did expose one missing feature of the sim, which is
weight sharing (e.g., for convolutions). In Loihi, it seems
that input and output axons can be shared for neurons.
Because of how kernels are used, we can massively
reduce the number of weights stored and instead link each
neuron to a kernel weight. We should be able to link multiple
neurons to one axon out i.e., when it has the same connectivity
to that core.

Because of the simplified implementation, the number of axons
used for DVS gesture is too large (both for inputs and outputs).
This won't affect the latency / energy estimates, only the
size of the axon map for DVS gesture.