Create initial version of Latin squares benchmark
Some hacky changes have been added but mostly this is a new
script that creates a Latin square of size NxN and uses
an SNN to create a solution. Although some tuning and playing
around is needed, it seems to work. The next step is to
port the network to Loihi and compare the energy and time
estimates.