Add basic parallelization using OpenMP
After profiling this doesn't really do much at the moment.
It makes a small difference in the message receiving pipeline.
Really, the majority of time is spent in the scheduling loop.

Really the focus needs to be optimizing the scheduler *if* I
want to improve performance (which isn't a focus right now).
Trying a heap instead of a list, or removing some
inefficiency in the loop so it doesn't spawn an event for
every neuron in every core (only for those that are
sending messages would be basic improvements).