Optimize message scheduler
Need to clean up a little better, but make sure now we don't
return to the priority queue for every neuron. We only need
to schedule the spiking neurons' spike messages. This results
in a significant speedup, since most neurons are not spiking.

The next worthwhile improvements could be to
* Parallelize the timing computations better. The only thing
   that must be serial is scheduling of messages, which is a
   fraction of that function. Move anything else outside e.g.,
   calculating the message timings.
* Use a heap based priority queue instead of a list