Add message trace feature
Not sure if the message struct should just be kept in each axon
map. That way, we'd avoid duplicating a bunch and as little
processing as possible would be kept in the serial portion of
code (the rest will be parallelized). Either way, there is some
trace capabilities for the message packets now. Later we can
move that struct around if needed.