Fix randomized and latin squares experiment
Make fixes to experimental scripts but also to SANA-FE to
accurately model Loihi for these two applications. Recalibrate
and improve the memory model slightly. With a sparse memory
representation the weight is more than 8-bits (we need to
also account for sign bits and compartment bits). This will
really affect the number of synapses per word and quite
significantly change the performance.

This does not affect DVS gesture which uses a different
synapse compression scheme. Now I am getting energy and
latency within about 10%, which seems like a reasonable
target.