Fix issues after merging experiments into main
Fix broken state of repo, need to merge in some of the changes
on the main branch. We need to manually save the mappings
of neurons to core.

Maybe I should reverse the flag - most of the time we want
to save the neuron mappings to cores. Only rarely we want
to suppress that feature e.g. for DVS