Fix sim.py positional args and forcing updates
Lots of small style cleanups and change how we "force" neurons
to update. Now, if a non-zero bias is detected we will update
the neuron. Otherwise, the "force" update option is a different
special case. Now all the experiments work for Loihi and
TrueNorth with no hacks. It also makes more sense, intuitively.