Try scheduling only messaging events
Reframe the scheduler - now we only schedule messaging events.
Its similar to before, and should produce the same time  and
behavior, but it is majorly refactored and should handle slightly
fewer events. Although only slightly optimised, this makes way
more sense when thinking about parallelizing the algorithm.
It's also simpler to explain at the high-level, but a TODO
is to properly comment this complex loop.