Clean up network metrics and neuron updating
Have a metric for all packet directions, north south east and
west. Clean up how we calculate the energy (just multiply counts
by cost per update at the end) and also simplify the active
/ inactive neuron update.

Next make the neuron updates additive. Have a baseline cost for
updating a neuron. If we receive spiking or bias inputs that's
an additional cost. If the neuron spikes and resets thats a third
additional cost.