Change how neuron update costs are tracked.
Instead of the Loihi specific active, inactive and spiking
costs. Have a cost for reading a neuron state, updating i.e.
writing a neuron state, and the cost of a spike output. Make
these costs additive. So now there's a baseline cost for all
neurons in the core, and if we update and write the potential
we add an extra update cost. If the neuron spikes, on top of
both costs we add another spike out cost.

Results are the same as before.