[wip] Schedule messages from a message buffer
Neuron simulation creates a buffer of
messages, which is then used by the scheduler.

This decouples the arch internals from the scheduler, which
should be standalone from the timestep loop.

Push for now, but redo some things before merging.

TODO: after this first implementation, it will be a lot cleaner if
we just have a timestep struct that contains a single message
buffer, and the simulation loop should just write directly to
this buffer. Remove the need to copy all messages over after
simulation finishes. Also, I want to rethink how we pass
timestep params over - maybe make the simulation struct for the
outer loop, and for the inner loop we just pass in the
timestep struct (instead of sim). Better separate what is needed
at each level.