[wip] Clean-up and simplify message buffering
Make changes so that we have a global buffer of messages in the
timestep struct (ts), and all neuron processing writes directly
into this buffer and queues in the struct. This simplified things
significantly, avoids redundant copying of data.

While we're at it, remove some redundant initialization and fix
small bugs where we access potentially uninitialized vars in
unused tiles and cores.