Continue NoC queueing model
Fix a number of bugs, including in how we calculate the packet
server rates, causing us to be orders of magnitude off. Also
move to a M/M/1/K/N model, which seems to be giving sane results
that make sense e.g., as we increase the buffer sizes, blocking
decreases and so wait time decreases across the chip.