[wip] Begin implementing event accurate NoC model
Still half completed. This model will be a highly detailed model
to roughly simulate all transactions in the NoC. While this is
very unlikely to be included in SANA-FE, now that I can calibrate
against real buffer sizes (and take a good guess at arbitration)
it will be a useful comparison point.