Finish implementing event-accurate NoC model
Redo the whole implementation to be much simpler, abeit slower since
every hop is now a new event. It was clear though that the complexity
was going to be unmanageable in the last approach. Now I model with
three events - send, hop and receive which can each spawn more events.
This leads to a reasonably simple implementation, using bits of the
older implementation. Now the code is running reliably even with
sanity checks.

Also do more queue analysis at scripts/queue_transient.py. Explore what
it will actually look like in the NoC buffers as packets arrive and
are processed.