Create new heuristic for network delay
Find some reasonable heuristic for network delay, that brings the
DVS gesture estimates into a reasonable range. This trades of
prediction accuracy for randomly generated vs DVS gesture, the
two stress tests.

The heuristic considers the messages along the links of the
NoC, rather than just the total messages at each router. The
direction matters it seems, in DVS gesture traffic is concentrated
along a few links, whereas in the random network activity is spread
across many hops and all the links, balancing the load. The heuristic
estimates delay as the sum of  messages per link multiplied by the
mean receive time in the NoC. Then, it seems importantly we scale
by the number of message hops. This scaling factor is not exact, but
it seems the more hops there are, the more we should scale down
the delay. Later we could tune things, like this delay scaling or the
fact that messages hold onto the receiving core even while they are
being delayed. It might make more sense to add a new event (like
attempted before), which tracks when messages are received and then
updates the receive timestamp.