Finish implementing more detailed scheduler
Messages are scheduled as being SENT and RECEIVED. Then we also
need to be careful when we model messages entering and leaving
the NoC. However, after experimenting, it seems that we still
need to adjust the network latency by some factor i.e., by the
number of message hops. After this adjustment, the accuracy
is roughly the same for all cases (slightly better for the
randomized SNNs).

Overall, although this implementation is a bit more intuitive.
I'm not sure (?) if the extra complexity is justified. We
don't majorly improve accuracy and now process 2x the number of
events. It seems fine to assume that messages arrive first-come
first-served from the sender.