[wip] Settle on heuristic for network delay
When calculating the messages in the NoC along a path, do not
include the link count going out from NoC to cores. However,
when updating links we divide by (hops+2), to account for that
link.

Excluding the NoC->core link seems to be more accurate for Luke's
network, and for the randomized networks. DVS gesture is about
24% accurate. Before, the assignment of weights to links was a
little arbitrary. This scheme is easier to explain and justify.
Still calculate the network delay using the average messages per
link. If I think of the messages per link as more of an abstract
weight per link (rather than actual messages buffered), then
I think it is easier to justify.