Merge C++ changes into latest main branch
Combine the C++ and plugin work with my changes to the scheduler,
and new tutorials. Revert a lot of the scheduler changes which for
some reason git decided to undo.

Also reformat and rename a bit to make consistent with the previous
code. That's just personal style though :)

Still some more integration work to do but this will be super
useful. Next will be converting the rest of the simulator to
be more C++. I can replace mallocs with vectors in a lot of places
which will improve the memory efficiency a lot and simplify some
things. Also use the priority STL class to implement the priority
queue. Overall, C++ should make the simulator more powerful.