[wip] Continue to develop c++ features
Convert the main simulation loop and kernel to be in C++ and use
C++ containers. Moving things to vector, map and queue will save
a lot of dead space and make the arch more flexible in the long
term. However, there are still a few things unimplemented here
and the simulator isn't quite working in its c++ form yet.