Disable openmp for now and remove unused variables
Disable OpenMP by default, this is heavy weight and not
really needed by the average user. But it requires extra
libs and dependencies. In the future, guard openmp by adding
another build flag which can enable it if needed.

Variables are only used when DEBUG is enabled. This causes a
clang warning/error.