sim.py: Add error messages for arch parsing
Add more helpful messages when parsing YAML file. This file is
likely to have user errors, so catch exceptions and print more
helpful debug. One error that is easy to make is mixing up the
arch/snn inputs. Check for the error that is likely when this
happens and print a recommendation.