Use std::filesystem and develop Python interface
Start to extend and improve the Python interface. Now the
Simulation object can return a RunData object - both after
running and also when the user requests a summary. The
RunData object contains all the summarized run statistics
that were / are stored in the run_summary.yaml

The next major interface improvement should allow a user to
build up SNNs and even architectures from Python. This should

* Allow users to create and save SNNs from Python. The C++ code
  should be able to dump the SNN to the custom .net format. The C++
  code can either load an SNN from a net file or have API calls
  that the Python can use (which will allow us to build many
  other network formats.

* Bypass the need for the messy architecture format. We can still
  support the old format for legacy reasons, but we should be able
  to use API calls to build the architecture from our YAML format.
  The SANA-FE library should officially just use and support this.