Replace Attribute struct with unordered_map
Use unordered_map<string, string> instead of custom defining my
own Attribute struct. This will be a lot more flexible when
integrating with Python, and allows me to leverage the STL.

The user passes attributes to architecture and network structs
using this mapping. For now, always use a string representation.
This enforces the representation of attributes as strings, so that
we know we can reliably save and load networks and architectures
to files. As soon as we allow the user to pass binary objects
around it will get messy / complicated.