Add NeuronGroup and Neuron python classes
Start binding more structs and functions across to Python.
Pybind11 does a good job of figuring out arguments and
return types. Now we can directly manipulate and build up
Networks inside of Python!

In the near future I should create functions to output the
contents of the structs for Python debugging. Also, in sim.py
create wrapper Classes, that can do a bit of extra parameter
checking and convert from generic dicts (e.g., containing a
mix of int, float and so on, that will convert to strings)