Add arch python interfaces
Expose the Architecture, Tile, Core and *Unit classes to Python.
This means creating classes and added functions to add new tiles,
cores etc. Now we should in theory be able to adapt the script
to build both Networks and Architectures without intermediate files

TODO:
* Extend the C++ kernel to dump the network and architecture
  files (raw arch and net formats).
* Add wrapper classes in the Python library and try implementing
  a network using this api.

The initialization has got slower again - maybe quickly profile
to see if we can gain back some speed when creating networks.
It's possible that moving to this new initialization code is a bit
slower, which may be a neccessary evil.