Efficiently support mapping in Python interface
Go back on implementation of attributes - only store the
attributes for the group once. This will be much more
efficient once we go to huge numbers of neurons. The
drawback is we need to store a pointer to the Network
in every neuron. This way, the neuron can get back to
its parent group.

To do this I assume the Network object and Architecture
objects can't be copied. This is probably safe enough for
how I intend to use these objects. They have too much state
built up to be copyable anyway...