Support setting of group-level params
After experimenting, it seems that setting up neurons and edges
individually is too slow - this needs to be done in the kernel.
Add support for setting group attributes and defining multiple
edges with a single Python API call.

In the process, I realised that the way I created the tile,
core and neuron group arrays was broken. Whenever the vector
was expanded it would reallocate and the Python object would
become garbage. If I want objects to be used more seamlessly,
I created a new slightly more complicated structure where we
have both a std::list, and a std::vector tracking references
to that same list for quick random access.

This should work well. Next is improving debugging and str
representation of the different structures.