Support formatting and prints for arch classes
Similarly support formatting of architecture classes.
Add a new print.cpp file. Now we can load and save both
network and architecture classes.

There is a remaining TODO to generate connected and
conv layer connectivity in the C++ code. The Python
should just be able to pass an iterable e.g., a Numpy
array, and it can generate the connections based on
weights and kernels. Doing this manually is very slow
in Python (especially for large fully connected networks).
This is lower priority.

Next up is adapting the various experimental scripts to
use the new API. Starting with the calibration i.e.,
power benchmark.