Debug arch saving and loading
Fix a bunch of bugs found by running larger arch through a load
and save. Now the save reproduces the loaded file mostly (just
reorders).

TODO: Use a map instead of unordered_map when printing attributes.
That way, they will be printed in sorted alphabetical order. It
would be nice if there is some predictable, determinstic order
to the attributes!