Change unordered_map->map
I don't really need the benefits of unordered map, and I've
realised that map keeps things in alphabetical order. That
makes for a more predictable and deterministic processing /
printing order when handling all the attributes. Also, map
may be better in the case where we iterate over all elements
anyway..