plugins.cpp: Support and re-enable soma plugins
Reintegrate the plugin mechanism, implemented by Ignacio
Gonzalez as part of the senior design capstone project.
Clean up this folder, add an automated build script and use
the Hodgkin-Huxley neuron as our example plugin (it's
unlikely to be useful as a h/w model so it's a good placeholder
example that we don't need to integrate into the kernel).

Now this plugin is integrated and working, with a few extra
debug prints left in for now.