[wip] Exploring pipeline implementations
Trying to clean up and simplify the pipeline model. Backtrack on
the dendrite implementation a bit - still need to figure this out.
Implement so that plugins just have to implement update() members
and get rid of input().

The idea is that update() routines should take a bool whether we
step time forward or not. Although I had this idea rejected before,
I believe it's required if we want to flexibly support all pipeline
configurations. For example, if the buffer is at the end of the pipeline
then all units can in theory receive multiple updates() before the
end of the time-step. If the buffer is before any of the units, they
receive only one update per time-step. Therefore, all models should
separate functionality that steps time forward e.g., leaks, which can
only happen once per timestep, and receiving spike-driven inputs,
which can happen multiple times per time-step.