Reformat to use spaces not tabs...
Not ideal at this stage of the project, but update my clang-tidy to
use spaces instead of tabs (and change the indent width to 4 spaces).
This will kill the ability to diff, but I figure we have rewritten
the whole codebase anyway (and I'm the only one to blame..)

The capstone project shows it will be harder to maintain tabs, whitespace will
eventually become messy with spaces creeping in. From this experiment, it seems
safer to just enforce spaces. 4 spaces seems to be more commonly accepted
viewing experience.