[wip] Begin rehaul of pipeline model
Start by rewriting the neuron processing loop. Next is message
processing.

Rethink and rewrite the pipeline model. The previous model recursion
which seemed like a good idea at the time, but proved impossible for
others to debug and develop. Reduce the caller depth and model the
pipeline as a small list of hardware units. Then iterate over the
units, hopefully leading to a more self documenting solution.
Functionally this will behave the same. Also, message processing is
TODO next.