Simplify pipeline model, regress back
Put more constraints on where the time-step buffer has to go
for this pipeline model, leading to required message processing
and neuron processing loops. If we want more complex pipelines,
we can implement new pipeline processing functions e.g., with
axon and synapse processing loops. However, a message processing
and neuron processing loop are the only *required* loops in all
models.

This simplifies the implementation somewhat and it looks like
a much older version of the code. At the same time, simplify
and clean-up the code to improve readability.