[wip] Redo and rethink how to implement dendrites
After discussion with the team, I'm changing the design so that the
hardware enforces the dendrite structure e.g., single compartment,
simple 1D tapped line, 2D branch structure or arbitrary graph
structure. It's up to the plugin to do that, and we'll instead set
attributes that control that e.g., individual tap parameters.

After other discussions, we will scrap the custom formats and move
back to parsing YAML (using an existing YAML parsing library) in
the kernel. That way, an attribute could be a more complex structure
like a list of values, allowing us to specify complex dendrite
configs using the network file. Similarly from Python, we just need
to pass in a dictionary, with basic types only.