[wip] Implementing neuron YAML parsing
Load the network file from YAML as well as the arch. This
required rethinking the network file and implementing more
YAML parsing code.

Tried a bunch of different options for implementing model
attributes / parameters so that we can decouple our plugins
from the YAML library. It seems that to generally support
different levels of nesting we will basically be (re)implementing
a YAML parsing structure. However, I think it's still useful to
separate our code from yaml-cpp, so fairly sure I'll stick with
NeuronAttribute2 (the more complex implementation).

Simpler alternatives include just restricting ourselves to one
layer of nesting. This removes the need for the Attribute
class (which supports casting to make things a lot more seamless
for the plugin writer). The parser would be similar, but
would no longer need recursive calls to itself. A more complex
version using template metaprogramming is also kept. However,
this still relies on undefined behavior, and so I think my version
with casting will be ok.

Currently in a state of mid-implementation, with the three
different Attribute attempts all present. I need to convert the
rest of the code, and rename NeuronAttribute to ModelParameter
or something...