Fix majority of MEDIUM or HIGH severity issues
From CodeChecker with all checkers enabled. A TODO will be to
go through each file in turn and get it down to 0 issues. I will
definitely need to disable a few checkers for the LOW severity
issues.

TODO: Also need to fix a couple of Coverity flagged issues.
Improving this can be an ongoing task, while improving the
parsing of the architecture and network files.