description.cpp: Add flexible range parsing
Sometimes (especially for network files) it would be nice to
support ID's that are just ranges e.g., g.0, g.1, without the
use of bracketed notation. Extend the range parsing a bit to
support this. Also support brackets that are used for a single
instance e.g., n[0]; make the ".." the delimiter when searching
for ranges rather than the open bracket '['.

Now ranges can either be specified between brackets e.g.,
n[0..2]
Which is expanded to n[0], n[1], n[2]. Or simply two integers e.g.,
0..2
Which is expanded to 0, 1, 2