[wip] Debug DVS gesture with new network format
Found a performance bug with the YAML usage and a parsing bug
for edges. It turns out that YAML-CPP is relatively slow for
very large (100MB+) files, but that should be okay.

It will be worth profiling performance to see if there are any
easy improvements to runtime, and especially memory usage which
seems too high (almost 8GB for processing a 100MB yaml file).
There should be easy optimizations to make.