Revert neurons to being stored in a vector
Instead of using a map container to store neurons, go back to using a vector.
When we create a neuron group, it must be a fixed size again. This solves
all the complication about the order of neurons being mapped etc. Each group
just has a fixed array of neurons. Most of the time, we'll probably just
allocate all neurons with the same parameters. Under the neuron section, it
is still possible to set the attributes of individual / ranges of neurons as
before.

Next improvement is to change how we store attributes. We can enable or
disable this feature and avoid duplication of attributes for soma/dendrite
etc. If we disable this, the user must map neurons before it sets their
attributes.