Fix edge attribute parsing
Fix when the edge attributes are given as a sequence e.g.,
0.0 -> 1.0 [weight=1]

This had temporarily been disabled. Temporarily fix this by
connecting neurons and then setting the edge attribute using
a reference to the edge. This will need to be modified as the
vector of connections can grow, invalidating any references.
Therefore, giving out a reference to this object is dangerous.

Alternatives are introducing an ConnectionTemplate containing all
the default attributes needed (this might be useful for a couple
of reasons). The function could also return an edge ID i.e., an
index into the vector, which will be safer if the list grows.

Improvements can be added later..