Support input nodes
Using the concept of models & plugins, it's easy to now define a
new input node class. For now, the input node can have a spike
train set, which allows the user to configure any sequence of
spikes. We could define more complex input generation in future,
for now this is sufficient.

In doing this, make some fixes to the vector / map casting of
model parameters. In doing this, I have some doubts in using
a variant. Either way, we may need to support more casting for
this model parameter class to be useful.