Extend how forcing updates works
Previously I had defined two force_update attributes, one in the
h/w description and one in the snn description. Confusingly they
got written to different variables and did different things.
Just have a single mechanism that we can force the soma, synapse
dendrite h/w to update. Change force_update to be specific to
each unit. If this is set in the neuron (or in future, in the h/w)
then by default, that unit must be updated every time-step.
Create force_soma_update, force_synapse_update, force_dendrite_update.

While doing this, clean up how we update synapses. I realised we
always pass in the address to the model, even though this information
isn't changing. We create a model for each synapse, so info such as
the synaptic address should just be passed once at creation.

In future, I might revert some of this so that we pass in nid, gid,
synapse id etc as attributes using the same mechanism that we pass
these in from the description formats. I just need to create code
that passes attributes from the simulator. I'd need to document
somewhere which attributes the simulator automatically sends, so that
plugin writers could figure how to handle these.