Remove redundant 'size' field in SNNs
In the new YAML format, the size field was added but is not really
needed, because we now have a compact way of specifying many neurons.
The group size can be figured out by the parser, so remove this
from the required fields, and add a function to calculate the total
neurons in the group.

TODO: add the old description format back for back-compability. This
will be slightly extended to allow for YAML-style attributes lists.