Default power metrics override power simulations
A bit of going in circles but develop the features for energy and
latency modelling further. Now the metrics provided in the arch
description can both act as a default option and also override
any custom user simulation. If neither model or metrics are
provided, throw an exception. The user can model energy, latency
both or neither in their implementations.

Simplify the abstract base class - now the user has just one
pure abstract update() they must implement. This seemed simpler.
Go back to a previous design where the user must return a special
Results object from update(), containing the simulated value
e.g., current, spike/no spike plus an additional, optional energy
and latency value. If the user wishes to model either of these
properties they must set the optional value and return it.

Pass the metrics as ModelParams to the H/W unit constructors
(inside the map<string,ModelParam>). This was done because the
energy and latency metrics are now optional (and we'll only catch
a missed metric once we get to simulating the energy and latency
and notice there's no metric set).

Depending on how the Python interface will do things, maybe revert
this back (and make energy & latency a std::optional<double> argument
to the constructor). On reflection, it might be clearer and more
explicit if these parameters are required by the constructor, rather
than hidden inside the model parameters again. Also, it allows for
the YAML parser to point to the location where things go wrong
e.g., if the user gives the wrong type or gives an incomplete set
of metrics.

TODO: cleanup and might be nice to have manual override for
turning on and off custom power simulation (for example, as a
fail-safe).