Explore simplifying h/w unit models
[WIP] Remove one level of indirection and remove the idea of models.
Instead the user will now implement classes for the hardware
units. We still need to move the Loihi model over, so that a single
unit supports multiple neurons. Then, we also need to support
instantiation of multiple units (e.g., for the multi-compartment
dendrite that supports a single neuron per dendrite).