Silence prints, fix bug with neuron address
The neuron address has to be per core, not per soma unit. This is
because we use the same address for the dendrite h/w.

If we do want to go back to tracking mapped soma / dendrite / synapse
id's, we will need to have separate variables for each. For now, just
track which order each neuron was mapped to the core.