Remove misleading pip and conda setups
These got committed but they're broken and out of date, so probably
just confusing.
TODO: actually setup a valid and minimal pip / conda env to use
sana-fe. This can be super minimal, we basically just need pybind11
to build sanafe. For running the experimental scripts we need libraries
such as numpy and matplotlib, so figure if we distribute this as another
conda environment...