[wip] Major rehaul of snn and arch structures.
Working with the simulator in Python showed that the current way
of modeling hardware is flaky. There are complex dependencies
between the architecture and the network that gets unreliable
when we want to dynamically build up objects e.g., in Python.

After a rethink, went back to the drawing board on how we
represent the classes. Properly manage ownership and avoid
having two interdependent objects. Now we have abstract
SpikingNetwork, and Architectures which describe attributes,
connectivity, and the architecture (abstract only). We then
build the SpikingHardware using the Architecture. Once we
built the hardware, we map the SpikingNetwork onto the
SpikingHardware.

The SpikingHardware manages everything, including all of its
MappedNeurons. SpikingHardware replaces Simulation and we can
now step simulation of each SpikingHardware object. This makes
the simulator much more robust e.g., for Python, with no global
state. It is much easier to dynamically change the SNN and
architecture without weird effects or dangling pointers. It's
also easy to copy networks and architectures from place to place.
It will in future also be easier to simulate multiple designs
in parallel.

Also, fix a couple of bugs introduced in the cpp branch.
Specifically, the hop energy/latency metrics were being copied
wrong. Also, the h/w lookup for soma and synapses was broken.
The soma h/w lookup was just picking the first unit and
now we make sure to look at the destination neuron's default
synapse h/w.