setup.py: More changes / fixes for python package
Currently, the setup.py does not copy __init__.py into the
package. When we install with pip we are missing this file
so the package won't work.

This might be an overly hacky way to get things all working. Maybe
need to bite the bullet and structure the directory a bit more
conventionally.