CMakeLists.txt: Make pybind11 locator more robust
Install was broken in some builds because CMake was unable to find
the pybind11 module. This was happening when running pip
install and setup.py, which creates a new venv for each build.

The fix is to add an extra conditional that can run Python and
print the pybind11 location. This seems to work on a variety of
platforms.

Also, fix how the parallel builds was being handled. It needs to
be set as an environment variable, not as a CMake variable.