Implement better tracing system
I've wanted better tracing for a while, I was previously  adding or
deleting prints when needed. Now support 4 trace levels which can
be customized per file (or category). This kind of solves the
problem where I want a lot of detail from some of the code
functionality, and minimal detail elsewhere.

Adding this new mechanism forces the compiler to look at all the
traces, even if the trace is disabled at build time. This is
better for maintainability but in this commit I fix a *lot* of
broken traces!

To use, pass flags to CMake e.g., -DDEBUG_LEVEL_SIM=3