chip.cpp: Change how we force updates, add resets
Forcing updates of hardware units in the receiving stage of the
pipeline was slightly wrong. It was causing us to update the
units a second time e.g., applying an extra leak term on the
dendrites before the soma would process the current. This was
causing results to mismatch.

Also extend the reset feature to work with the dendrite h/w, and
make sure to reset any temporary pipeline state.