[wip] Most recent experimental updates
Commit temporary / hacky code exploring the timing scheduling,
including a simplified timing model. This needs to be cleaned up
a bit in the future, but commit to capture the state of the repo
(nothing has been broken by this).

Minor updates to LASAGNA (LASANA?) script for modeling analog
neurons, those experiments are captured in another private
repo for now...