Extend and use generic PipelineUnit class
Instead of having specialized unit classes i.e., SynapseUnit,
DendriteUnit, SomaUnit, just simplify and go with the single
superclass: PipelineUnit. This can support a pipeline hardware
unit that supports one or any combination of the Synapse, Dendrite
or Soma functionality. The user must then implement the
appropriate update() functionality. All units contain all the
possible counters.

In some ways this is simpler and "cleaner" than manually specifying
each unit and each possible unit combination. However, the interface
is a little less concrete, as the user has to figure out which
update() to implement and which return value. By improving the user
feedback though, this shouldn't be too difficult for the user.

TODO: clean up / refactor the new process_connection and
process_neuron member functions. These are too long and convoluted.
Come up with shorter and cleaner routines that are clearer to the
user.