[TCAD REBUTTAL] Finalize experiments and plots
Mostly just adjusting different plots to account for the TCAD
rebuttal, addressing reviewer's feedback. The most significant
change was addressing the point of "why" use a more complex
network model. To do this, we collected data on the "analytical"
model or old/simple latency model, which is assuming a simple
fixed network hop cost and ideal communication